Back end/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pymysql
passlib[bcrypt]
pyjwt
python-dotenv
cryptography
